+++ packages/eas-cli/src/graphql/types/Update.ts
@@ -9,6 +9,7 @@ export const UpdateFragmentNode = gql`
     runtimeVersion
     platform
     manifestFragment
+    isRollBackToEmbedded
     manifestPermalink
     gitCommitHash
     actor {